[FIX] purchase_stock: _run_buy check supplier qty
Usecase to reproduce:
- Product with routes MTO and BUY
- Add a supplier on a product with a min qty of 5
- Create a SO with a qty of 1

A PO is created with the supplier and a quantity of 1. It should return
an error since there is no suppliers that sell with this quantity.

It happens because _run_buy directly use the suppliers on the product
without check for date and minimal quantity.
This commit replace it by _select_seller that manage every details.