[FIX] sale: fix wording of link when sending a quotation by email
Since commit be00baedd03ccaeb9a9f37bfce0b6511c2992b27 wording of link when
sending a quotation by email is broken. Indeed has document in draft is
not correctly considered as having to be signed and/or paid the link is
not labelled correctly.

This commit fixes it. Parameter is also renamed to be more understandable.
This commit is linked to task ID 1910532 and PR #29849 .

closes odoo/odoo#30953