[FIX] account: remove child taxes if the tax is not a group
Create a tax of type "group of taxes".
Put taxes in its children_tax_ids field.
Change the type to something else.
The view hides the field children_tax_ids, since it doesn't make sense anymore.
However it doesn't actually remove the children_tax_ids.
As a result the child taxes are use in the computation.

We add on onchange to avoid that unfortunate situation.

opw 1931089

closes odoo/odoo#30727