[IMP] base: improve search on res.partner tags to include parent tags
When using the 'tag' quicksearch on the contacts, it should search on the
parent tags as well. It is what users expect from this kind of search.

Using child_of is considered valid here as people should not have a huge
list of nested categories / tags. Even a famous production db like Odoo
does not have more than 215 of them.

Related Task ID: 1903699
closes #30733

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>