[FIX] mail: mention special char=>link partner
This is a regression from: 3a17d13d2. In it the visual side was solved,
but unexpectedly the functional side still worked on unescaped content.

Thus visually a @partnèr could be seen as a link, but it was no anymore
added as follower.

Without change, modified test fails with:

 mentioned partners are sent to server (expected: [1], result: [])

note: change is partial backport of 12.2 cb5afdfee

opw-1931247
closes #30846